webcrawl/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
confection==0.1.5
contourpy==1.3.1
cryptography==43.0.3
curated-tokenizers==0.0.9
curated-transformers==0.1.1
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.10
darkdetect==0.8.0
debugpy==1.8.9
decorator==5.1.1
diffusers==0.30.3
dnspython==2.7.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
executing==2.1.0
fastapi==0.112.2
fastapi-cli==0.0.5
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.9.0
ftfy==6.2.3
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
gTTS==2.5.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.24.7
ibm-cloud-sdk-core==3.20.6
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-watson==8.1.0
ibm_watsonx_ai==1.1.7
idna==3.10
importlib_metadata==8.5.0
invisible-watermark==0.2.0
ipykernel==6.29.5
ipython==8.29.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
langcodes==3.5.0
langid==1.1.6
language_data==1.3.0
libclang==18.1.1
libretranslatepy==2.1.1
lomond==0.3.3
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.4.1
mpmath==1.3.0
murmurhash==1.0.11
namex==0.0.8
narwhals==1.8.1
nest-asyncio==1.6.0
networkx==3.3
nodemon==0.0.16
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.1.4
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
posenet==0.0.3
preshed==3.0.9
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.8.0
PyMySQL==1.1.1
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-version==0.0.2
pyttsx3==2.98
pytz==2024.2
PyWavelets==1.7.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.2
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spacy==3.8.2
spacy-alignments==0.9.1
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
srsly==2.4.8
stack-data==0.6.3
starlette==0.38.6
streamlit==1.38.0
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
tf_keras==2.18.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.15.2
toml==0.10.2
torch==2.4.1+cu118
torchaudio==2.4.1
torchvision==0.19.1+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.36.2
translate==3.6.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
uvicorn==0.32.0
wasabi==1.1.3
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
